# specs/001-scoping-this-feature/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: AI Flight Search Chat API
  description: API for AI-powered conversational flight search system
  version: 1.0.0
  contact:
    name: Paylater Travel
    url: https://paylatertravel.com.au
servers:
  - url: https://api.paylatertravel.com.au/v1
    description: Production server
  - url: http://localhost:3000/api
    description: Development server
paths:
...
  /destinations/recommendations:
    get:
      summary: Get destination recommendations
      description: Returns categorized destination suggestions for guided discovery
      operationId: getDestinationRecommendations
      parameters:
        - name: category
          in: query
          schema:
            type: string
            enum: [island-vibes, mountain-views, snowy-adventures, city-escapes, wine-tours, none-of-these]
          description: Filter recommendations by category
      responses:
        '200':
          description: Recommendations retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  categories:
                    type: array
                    items:
                      type: object
                      properties:
                        category:
                          type: string
                          example: "island-vibes"
                        display_name:
                          type: string
                          example: "Island vibes"
                        destinations:
                          type: array
                          items:
                            $ref: '#/components/schemas/DestinationRecommendation'
...
components:
  schemas:
...
    DestinationRecommendation:
      type: object
      properties:
        id:
          type: string
          format: uuid
        category:
          type: string
          example: "island-vibes"
        category_display_name:
          type: string
          example: "Island vibes"
        name:
          type: string
          example: "Fiji"
        iata_code:
          type: string
          pattern: "^[A-Z]{3}$"
          example: "NAN"
        description:
          type: string
          example: "Tropical paradise with crystal clear waters and friendly locals"
        image_url:
          type: string
          format: uri
          example: "https://cdn.paylatertravel.com/destinations/fiji.jpg"
        display_order:
          type: integer
          example: 1
        is_active:
          type: boolean
          example: true
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
security:
  - ApiKeyAuth: []
...